+++ modules/template/elements/SpacesElementVariable.php
@@ -10,16 +10,12 @@ namespace humhub\modules\custom_pages\modules\template\elements;
 
 class SpacesElementVariable extends BaseElementVariableIterator
 {
-
     public function __construct(SpacesElement $elementContent)
     {
         parent::__construct($elementContent);
 
         foreach ($elementContent->getItems() as $space) {
-            $spaceVariable = new SpaceElementVariable($elementContent);
-            $spaceVariable->setContentContainer($space);
-
-            $this->items[] = $spaceVariable;
+            $this->items[] = SpaceElementVariable::instance($elementContent)->setRecord($space);
         }
     }
 }