+++ tests/codeception/acceptance/CreatePhpPageCest.php
@@ -52,7 +52,7 @@ class CreatePhpPageCest
         $I->jsShow('.form-collapsible-fields.closed fieldset');
         $I->fillField('CustomPage[sort_order]', '400');
         $I->selectOption('CustomPage[icon]', ['value' => 'fa-adn']);
-        $I->click('Save');
+        $I->click('Create');
         $I->waitForElementVisible('#topbar-second .fa-adn');
         $I->expectTo('see my new page in the top navigation');